fix: use fetch
