+++ package.json
@@ -13,9 +13,8 @@
   "license": "MIT",
   "dependencies": {
     "@logseq/libs": "0.0.6",
-    "@sentry/react": "^6.19.7",
-    "@sentry/tracing": "^6.19.7",
-    "axios": "^0.27.2",
+    "@sentry/react": "^7.1.1",
+    "@sentry/tracing": "^7.1.1",
     "cheerio": "^1.0.0-rc.11",
     "firebase": "^9.8.3",
     "hash.js": "^1.1.7",
@@ -28,28 +27,29 @@
     "@semantic-release/changelog": "6.0.1",
     "@semantic-release/exec": "6.0.3",
     "@semantic-release/git": "10.0.1",
-    "@semantic-release/npm": "9.0.0",
-    "@sentry/cli": "^1.73.1",
-    "@types/jest": "^27.5.1",
+    "@semantic-release/npm": "9.0.1",
+    "@sentry/cli": "^2.1.0",
+    "@types/jest": "^28.1.1",
     "@types/luxon": "^2.3.0",
-    "@types/node": "17.0.1",
-    "@types/react": "17.0.39",
-    "@types/react-dom": "17.0.11",
-    "@typescript-eslint/eslint-plugin": "5.12.1",
-    "@typescript-eslint/parser": "5.12.1",
-    "@vitejs/plugin-react": "1.2.0",
-    "conventional-changelog-conventionalcommits": "4.6.3",
-    "eslint": "8.10.0",
-    "eslint-plugin-react": "7.29.2",
-    "jest": "^28.1.0",
+    "@types/node": "18.0.0",
+    "@types/react": "18.0.13",
+    "@types/react-dom": "18.0.5",
+    "@typescript-eslint/eslint-plugin": "5.28.0",
+    "@typescript-eslint/parser": "5.28.0",
+    "@vitejs/plugin-react": "1.3.2",
+    "axios": "^0.27.2",
+    "conventional-changelog-conventionalcommits": "5.0.0",
+    "eslint": "8.17.0",
+    "eslint-plugin-react": "7.30.0",
+    "jest": "^28.1.1",
     "jest-environment-jsdom": "^28.1.1",
-    "semantic-release": "19.0.2",
-    "ts-jest": "^28.0.3",
-    "typescript": "4.5.5",
-    "vite": "2.8.4",
+    "semantic-release": "19.0.3",
+    "ts-jest": "^28.0.5",
+    "typescript": "4.7.3",
+    "vite": "2.9.12",
     "vite-plugin-react-svg": "^0.2.0",
-    "vite-plugin-windicss": "1.8.1",
-    "windicss": "3.5.0"
+    "vite-plugin-windicss": "1.8.4",
+    "windicss": "3.5.4"
   },
   "logseq": {
     "id": "logseq-my-highlights",
