+++ src/hooks/useBookDetailsSearch.tsx
@@ -1,32 +1,22 @@
 import React from 'react';
-import Axios from 'axios';
-import * as cheerio from 'cheerio';
+import { AmazonSearchResult, parseAmazonSearchResults } from '../utils/parseAmazonSearchResults';
 
-export interface DetailsSearchResult {
-  asin: string;
-  title: string;
-  imageUrl: string;
-  author?: string;
-  published?: string;
-  productPath: string;
-}
 
-const searchAmazonKindle = async (query: string): Promise<Array<DetailsSearchResult>> => {
+const searchAmazonKindle = async (query: string): Promise<Array<AmazonSearchResult>> => {
   const SEARCH_URL = "https://www.amazon.com/s";
-  let results = [] as Array<DetailsSearchResult>;
-  const params = {
-    i: 'digital-text',
-    k: query,
-  };
+  let results = [] as Array<AmazonSearchResult>;
 
+  const url = new URL(SEARCH_URL);
+  url.searchParams.append('i', 'digital-text');
+  url.searchParams.append('k', query);
 
-  const response = await Axios.get(SEARCH_URL, {
-    params,
-    validateStatus: () => true,
+  const response = await fetch(url.toString(), {
+    mode: 'no-cors'
   });
 
   if (response.status === 200) {
-    results = parseSearchResults(response.data);
+    const content = await response.text();
+    results = parseAmazonSearchResults(content);
   }
 
   return results;
@@ -37,7 +27,7 @@ interface UseBookDetailsSearchParams {
 }
 
 export const useBookDetailsSearch = () => {
-  const [results, setResults] = React.useState<Array<DetailsSearchResult>>([]);
+  const [results, setResults] = React.useState<Array<AmazonSearchResult>>([]);
   const [isBusy, setBusy] = React.useState(false);
 
   const search = async (query?: string) => {
@@ -57,34 +47,4 @@ export const useBookDetailsSearch = () => {
 }
 
 
-export function parseSearchResults(content: string): DetailsSearchResult[] {
-  const results = [] as DetailsSearchResult[];
-  const $ = cheerio.load(content);
 
-  const resultContainer = $('.s-main-slot').first();
-
-  resultContainer.children('div[data-component-type=s-search-result]').each((i, elem) => {
-    const e = $(elem);
-    const asin = e.attr('data-asin');
-
-    const imageRecord = e.find('img[class=s-image]').first();
-    const title = imageRecord.attr('alt');
-    const imageUrl = imageRecord.attr('src');
-
-    const imageContainer = e.find('span[data-component-type=s-product-image]').first();
-    const productLink = imageContainer.find('a').first();
-    let productPath = productLink.attr('href');
-    productPath = productPath?.substring(0, productPath.lastIndexOf('/'))
-
-    const headerElement = e.find('.s-title-instructions-style').first();
-    const metaElements = $(headerElement[0].children[1]).find('.a-row').first();
-    const author = $(metaElements).find('.s-link-style.a-size-base').first().text();
-    const published = $(metaElements).find('.a-text-normal').last().text();
-
-    if (asin && title && imageUrl && productPath && !title.startsWith('Sponsored Ad')) {
-      results.push({ asin, title, imageUrl, author, published, productPath });
-    }
-  });
-
-  return results;
-}
