+++ src/utils/parseAmazonSearchResults.test.ts
@@ -1,8 +1,7 @@
-import { parseSearchResults } from "./useBookDetailsSearch";
+import { parseAmazonSearchResults } from "./parseAmazonSearchResults";
 import * as fs from 'fs';
 import * as path from 'path';
 
-describe('useBookDetailsSearch', () => {
   describe('parseSearchResults', () => {
     let content: string;
     beforeAll(async () => {
@@ -12,32 +11,31 @@ describe('useBookDetailsSearch', () => {
     });
 
     it('should parse all results', () => {
-      const results = parseSearchResults(content);
+      const results = parseAmazonSearchResults(content);
       expect(results).toHaveLength(16);
     });
 
     it('should parse ASIN', () => {
-      const results = parseSearchResults(content);
+      const results = parseAmazonSearchResults(content);
       const first = results[0];
       expect(first).toHaveProperty('asin', 'B09V5M8FR5');
     });
 
     it('should parse author', () => {
-      const results = parseSearchResults(content);
+      const results = parseAmazonSearchResults(content);
       const first = results[0];
       expect(first).toHaveProperty('author', 'Sönke Ahrens');
     });
 
     it('should parse title', () => {
-      const results = parseSearchResults(content);
+      const results = parseAmazonSearchResults(content);
       const first = results[0];
       expect(first).toHaveProperty('title', 'How to Take Smart Notes: One Simple Technique to Boost Writing, Learning and Thinking');
     });
 
     it('should parse product path', () => {
-      const results = parseSearchResults(content);
+      const results = parseAmazonSearchResults(content);
       const first = results[0];
       expect(first).toHaveProperty('productPath', '/How-Take-Smart-Notes-Technique-ebook/dp/B09V5M8FR5');
     });
   });
-});
